requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2024.7.1
django-cors-headers==4.4.0
django-redis==5.4.0
celery==5.4.0
redis==5.2.1
sentry-sdk==2.19.2
psycopg2-binary==2.9.9
gunicorn==23.0.0

crawl4ai>=0.7.0
playwright>=1.47

httpx>=0.27
lxml>=5.3
tqdm>=4.66
beautifulsoup4>=4.12
requests>=2.32
